WIP: Error refactoring #6
