+++ smb/Cargo.toml
@@ -23,6 +23,7 @@ maybe-async = "0.2"
 rand = "0.8.5"
 log = "0.4.22"
 time = {version = "0.3.37", features = ["macros"]}
+thiserror = "2.0"
 
 # APIs
 sspi = { version = "0.15.1" }
@@ -32,10 +33,12 @@ gss-api = {version = "0.1.0", features = []}
 hmac = "0.12.1"
 sha2 = "0.10.8"
 rust-kbkdf = {version = "1.0.1"}
+crypto-common = {version = "0.1"}
 aes = "0.8.4"
 aes-gcm = {version = "0.10.3", features = ["std"], optional = true}
 cmac = {version = "0.7.2", optional = true}
 ccm = {version = "0.5.0", optional = true}
+aead = {version = "0.5", optional = true}
 
 # Compression
 lz4_flex = { version = "0.11", default-features = false, features = ["std"], optional = true }
@@ -53,8 +56,8 @@ compress_lz4 = ["dep:lz4_flex"]
 compress = ["compress_pattern_v1", "compress_lz4"]
 
 # Encryption
-encrypt_aes128ccm = ["dep:ccm"]
-encrypt_aes256ccm = ["dep:ccm"]
+encrypt_aes128ccm = ["dep:ccm", "dep:aead"]
+encrypt_aes256ccm = ["dep:ccm", "dep:aead"]
 encrypt_aesccm = ["encrypt_aes128ccm", "encrypt_aes256ccm"]
 encrypt = ["encrypt_aesccm"]
