+++ smb/src/connection/netbios_client.rs
@@ -3,12 +3,12 @@ use std::io::Cursor;
 
 #[cfg(not(feature = "async"))]
 use std::{
-    io::{Read, Write},
+    io::{self, Read, Write},
     net::TcpStream,
 };
 #[cfg(feature = "async")]
 use tokio::{
-    io::{AsyncReadExt, AsyncWriteExt},
+    io::{self, AsyncReadExt, AsyncWriteExt},
     net::TcpStream,
 };
 
@@ -16,11 +16,13 @@ use binrw::prelude::*;
 
 use crate::packets::netbios::{NetBiosMessageContent, NetBiosTcpMessage, NetBiosTcpMessageHeader};
 
+use super::Error;
+
 /// A (very) simple NETBIOS client.
 ///
 /// This client is NOT thread-safe, and should only be used for SMB wraaping.
-/// 
-/// Use [connect](NetBiosClient::connect), [send](NetBiosClient::send), 
+///
+/// Use [connect](NetBiosClient::connect), [send](NetBiosClient::send),
 /// and [receive_bytes](NetBiosClient::recieve_bytes) to interact with a server.
 pub struct NetBiosClient {
     connection: Option<TcpStream>,
@@ -33,31 +35,25 @@ impl NetBiosClient {
 
     /// Connects to a NetBios server in the specified address.
     #[maybe_async]
-    pub async fn connect(&mut self, address: &str) -> Result<(), Box<dyn std::error::Error>> {
+    pub async fn connect(&mut self, address: &str) -> Result<(), io::Error> {
         self.connection = Some(TcpStream::connect(address).await?);
         Ok(())
     }
 
     /// Sends a NetBios message.
     #[maybe_async]
-    pub async fn send(
-        &mut self,
-        data: NetBiosMessageContent,
-    ) -> Result<(), Box<dyn std::error::Error>> {
+    pub async fn send(&mut self, data: NetBiosMessageContent) -> Result<(), crate::Error> {
         let raw_message = NetBiosTcpMessage::from_content(&data)?;
-        self.send_raw(raw_message).await
+        Ok(self.send_raw(raw_message).await?)
     }
 
     /// Sends a raw byte array of a NetBios message.
     #[maybe_async]
-    pub async fn send_raw(
-        &mut self,
-        data: NetBiosTcpMessage,
-    ) -> Result<(), Box<dyn std::error::Error>> {
+    pub async fn send_raw(&mut self, data: NetBiosTcpMessage) -> Result<(), crate::Error> {
         // TODO(?): assert data is a valid and not-too-large NetBios message.
         self.connection
             .as_mut()
-            .ok_or("NetBiosClient is not connected")?
+            .ok_or(crate::Error::NotConnectedError)?
             .write_all(&data.to_bytes()?)
             .await?;
 
@@ -66,11 +62,8 @@ impl NetBiosClient {
 
     // Recieves and parses a NetBios message header, without parsing the message data.
     #[maybe_async]
-    pub async fn recieve_bytes(&mut self) -> Result<NetBiosTcpMessage, Box<dyn std::error::Error>> {
-        let tcp = self
-            .connection
-            .as_mut()
-            .ok_or("NetBiosClient is not connected")?;
+    pub async fn recieve_bytes(&mut self) -> Result<NetBiosTcpMessage, crate::Error> {
+        let tcp = self.connection.as_mut().ok_or(crate::Error::NotConnectedError)?;
 
         // Recieve header.
         let mut header_data = vec![0; NetBiosTcpMessageHeader::SIZE];
@@ -78,7 +71,7 @@ impl NetBiosClient {
         let header = NetBiosTcpMessageHeader::read(&mut Cursor::new(header_data))?;
 
         if header.stream_protocol_length.value > 2u32.pow(3 * 8) - 1 {
-            return Err("Stream protocol length is too large".into());
+            return Err(crate::Error::InvalidMessage("Message too large.".into()));
         }
 
         // Recieve message data.
