+++ smb/src/crypto.rs
@@ -5,3 +5,16 @@ mod signing;
 pub use encryption::{make_encrypting_algo, EncryptingAlgo, ENCRYPTING_ALGOS};
 pub use kbkdf::{kbkdf_hmacsha256, DerivedKey, KeyToDerive};
 pub use signing::{make_signing_algo, SigningAlgo, SIGNING_ALGOS};
+
+use thiserror::Error;
+use crypto_common::InvalidLength;
+
+#[derive(Debug, Error)]
+pub enum CryptoError {
+    #[error("Invalid length")]
+    InvalidLength(#[from] InvalidLength),
+    #[error("Unsupported algorithm")]
+    UnsupportedAlgorithm,
+    #[error("AEAD calculation error")]
+    AeadError(#[from] aead::Error),
+}
